pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "python-analytics-accelerator"
version = "0.2.0"
authors = [{ name = "Cody", email = "cody@dkdc.dev" }]
description = "Python analytics solution accelerator using composable data systems"
readme = "readme.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'gcsfs',
    'ibis-framework[duckdb,clickhouse,deltalake]',
    'typer',
    'python-dotenv',
    'plotly',
    'streamlit',
    'great-tables',
    'dagster',
    'dagster-webserver',
]

[project.urls]
"Homepage" = "https://github.com/lostmygithubaccount/python-analytics-accelerator"
"Bug Tracker" = "https://github.com/lostmygithubaccount/python-analytics-accelerator/issues"

[project.scripts]
acc = "python_analytics_accelerator.cli:app"

[tool.ruff]
extend-include = ["*.ipynb"]
